% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable_track_changes.R
\name{enable_track_changes}
\alias{enable_track_changes}
\title{Enable css for track changes in knitted rmarkdown}
\usage{
enable_track_changes()
}
\value{
CSS and javascript text snippet
}
\description{
Enable css for track changes in knitted rmarkdown
}
