% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_changes.R
\name{track_changes}
\alias{track_changes}
\title{Start tracking changes on a file}
\usage{
track_changes(path, wait = 1)
}
\arguments{
\item{path}{The path to the file to track.}

\item{wait}{The number of seconds between each check for changes}
}
\description{
Start tracking all changes to a file and preview changes in the viewer.
The changes will be added to the original file as markup when done.
}
\examples{
trackChanges("my_file.txt")

}
