% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackchanger.R
\name{trackchanger}
\alias{trackchanger}
\title{Shiny Gadget to interactively accept or reject chaanges in rmarkdown}
\usage{
trackchanger()
}
\description{
Open a rmd document, then launch this addin via RStudio menu.
It will propose to accept or reject the change or skip to the next function.
}
\examples{
\dontrun{

trackchanger()

}
}
